# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ord",
    "description": "◉ Ordinal wallet and block explorer",
    "license": {
      "name": "CC0-1.0"
    },
    "version": "0.14.1"
  },
  "paths": {
    "/api/v1/brc20/address/{address}/balance": {
      "get": {
        "tags": [
          "brc20"
        ],
        "summary": "Get all ticker balances of the address.",
        "description": "Get all ticker balances of the address.\n\nRetrieve all BRC20 protocol asset balances associated with a address.",
        "operationId": "brc20_all_balance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account balances by query address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBRC20AllBalance"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20/tick/{ticker}/address/{address}/balance": {
      "get": {
        "tags": [
          "brc20"
        ],
        "summary": "Get the ticker balance of the address.",
        "description": "Get the ticker balance of the address.\n\nRetrieve the asset balance of the 'ticker' for the address.",
        "operationId": "brc20_balance",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "Token ticker",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 4,
              "minLength": 4
            }
          },
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account balance by query ticker.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBRC20Balance"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiBRC20AllBalance": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.ApiBalances"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ApiBRC20Balance": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.ApiBalance"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string",
                "description": "Internal server error."
              }
            },
            "example": {
              "code": 1,
              "msg": "internal error"
            }
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "Bad request."
              }
            },
            "example": {
              "code": 1,
              "msg": "bad request"
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "Resource not found."
              }
            },
            "example": {
              "code": 1,
              "msg": "not found"
            }
          }
        ]
      },
...
      "brc20.ApiBalance": {
        "type": "object",
        "required": [
          "tick",
          "availableBalance",
          "transferableBalance",
          "overallBalance"
        ],
        "properties": {
          "availableBalance": {
            "type": "string",
            "format": "uint64",
            "description": "Available balance."
          },
          "overallBalance": {
            "type": "string",
            "format": "uint64",
            "description": "Overall balance."
          },
          "tick": {
            "type": "string",
            "description": "Name of the ticker."
          },
          "transferableBalance": {
            "type": "string",
            "format": "uint64",
            "description": "Transferable balance."
          }
        }
      },
      "brc20.ApiBalances": {
        "type": "object",
        "required": [
          "balance"
        ],
        "properties": {
          "balance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20.ApiBalance"
            }
          }
        }
      },
...
    }
  }
}